feat(lc-dapp): disconnect provider when chain id is changed
