+++ examples/light-client-dapp/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./ConnectedAccount"
 export * from "./Transfer"
+export * from "./ChainSelect"
