+++ examples/light-client-dapp/src/settings.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_CHAIN_ID =
+  "0xe143f23803ac50e8f6f8e62695d1ce9e4e1d68aa36c1cd2cfd15340213f3423e" // westend
